--- demo/quicksort/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "quicksort"
-version = "0.1.0"
-authors = ["Niko Matsakis <niko@alum.mit.edu>"]
-
-[dependencies]
-rand = "0.3"
-rayon = { path = "../../" }
-time = "0.1"
